% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.functions.R, R/Survey.R
\name{analyse.data}
\alias{analyse.data}
\alias{analyse.data,Survey,DS.Analysis-method}
\title{S4 generic method to run analyses}
\usage{
analyse.data(survey, analysis, ...)

\S4method{analyse.data}{Survey,DS.Analysis}(survey, analysis, warnings = list())
}
\arguments{
\item{analysis}{an object of class DS.Analysis}

\item{...}{optional arguments including the following:}

\item{warnings}{a list of warnings and how many times they arose}

\item{object}{an object of class Survey}

\item{point}{logical indicating whether it is a point transect survey}
}
\value{
a list containing an S3 ddf object and optionally an S3 dht object relating to the model with the minimum criteria.
}
\description{
This method carries out an analysis of distance sampling data. This method
is provided to allow the user to perform diagnostics of the analyses used
in the simulation. The data argument can be obtained by a call to
\code{simulate.survey(object, dht.table = TRUE)}. Note if the first object
supplied is of class DDf.Analysis then the second argument must be of class
DDf.Data. The data argument may be of either class for an object argument
of class Simulation.
}
