% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.Summary.R
\docType{class}
\name{Density.Summary-class}
\alias{Density.Summary-class}
\title{Class "Density.Summary"}
\description{
Class \code{"Density.Summary"} is an S4 class containing a
summary of the density grids for each strata.
}
\section{Slots}{

\describe{
\item{\code{summary}}{a summary of the average abundances and densities for
each strata.}
}}

\seealso{
\code{\link{make.density}}
}
\keyword{classes}
