% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{generate.transects,Simulation-method}
\alias{generate.transects,Simulation-method}
\title{generate.transects}
\usage{
\S4method{generate.transects}{Simulation}(object, quiet = FALSE, ...)
}
\arguments{
\item{object}{object of class Simulation}

\item{quiet}{if TRUE silences some warnings}

\item{...}{not implemented}
}
\value{
an object of class Transect from dssd package
}
\description{
Generates a set of transects based on the design provided.
}
