% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{get.densities}
\alias{get.densities}
\title{Method to get density values}
\usage{
get.densities(density, coords = FALSE)
}
\arguments{
\item{density}{object of class Density}

\item{coords}{if TRUE also returns x, y coordinates}
}
\value{
either returns a numeric vector of density values or a dataframe
with columns x, y and density.
}
\description{
This method extracts the density values from a density object. It will
optionally also return the x and y centre points for the density grid
cells.
}
