% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.functions.R
\name{get.distance.data}
\alias{get.distance.data}
\title{S4 generic method to extract distance data}
\usage{
get.distance.data(object)
}
\arguments{
\item{object}{an object of class LT.Survey.Results}
}
\value{
a list containing a data.frame of distance data and a vector of all
the possible detection distances. The latter can be used to calculate the
number of animals / clusters in the covered area for truncation distances
equal to or less than those used in the design.
}
\description{
Extracts distance data from a Survey.Results object
}
\seealso{
\code{\link{create.survey.results}}
}
