% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{histogram.N.ests}
\alias{histogram.N.ests}
\title{histogram.N.ests}
\usage{
histogram.N.ests(x, ...)
}
\arguments{
\item{x}{object of class Simulation}

\item{...}{optional parameters to pass to the generic hist function in graphics}
}
\value{
No return value, displays a histogram of the abundance estimates
}
\description{
Plots a histogram of the estimates abundances
}
