% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassConstructors.R
\name{make.density}
\alias{make.density}
\title{Creates a Density object}
\usage{
make.density(
  region = make.region(),
  density.surface = list(),
  x.space = 5,
  y.space = NULL,
  buffer = numeric(0),
  constant = numeric(0),
  density.gam = NULL
)
}
\arguments{
\item{region}{the Region object in which the density grid will be created}

\item{density.surface}{Object of class \code{list}; list of
data.frames with the columns x, y and density. There must be one
data.frame for each strata.}

\item{x.space}{the intervals in the grid in the x direction}

\item{y.space}{the intervals in the grid in the y direction}

\item{buffer}{the width of the buffer region for generating the density grid. If not supplied DSsim will use the maximum value provided for the x.space or y.space.}

\item{constant}{a value describing a constant density across the surface. If not supplied a default value of 1 is used for all strata.}

\item{density.gam}{\code{gam} object created using \code{mgcv} with only x and y as explanatory covariates.}
}
\value{
object of class Density
}
\description{
The user has the option to create a grid describing the density of the
objects and pass this in giving the x and y spacings used in the creation
of this grid. Alternatively the user can specify a constant density and x,
y spacings and this grid will be generated automatically. The user may
also supply a \code{mgcv gam} object and x, y spacings and the density grid will
be created from these.
}
\examples{
# A simple density surface with a constant value of 1 can be created within a rectangular
# region using
# the default values:
density <- make.density(re)
plot(density)

# The example below shows hot to add high and low point to the density surface
region <- make.region()
pop.density <- make.density(region = region, x.space = 10,
 y.space = 10, constant = 0.5)

pop.density <- add.hotspot(pop.density, centre = c(50, 200),
 sigma = 100, amplitude = 0.1)
pop.density <- add.hotspot(pop.density, centre = c(500, 700),
 sigma = 900, amplitude = 0.05)
pop.density <- add.hotspot(pop.density, centre = c(300, 100),
 sigma = 100, amplitude = -0.15)

#New plot features
plot(pop.density)

#Block style plotting
plot(pop.density, contours = FALSE, style = "blocks")

}
\seealso{
\code{\link{make.region}}
}
\author{
Laura Marshall
}
