% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassConstructors.R
\name{make.detectability}
\alias{make.detectability}
\title{Creates a Detectablility object}
\usage{
make.detectability(
  key.function = "hn",
  scale.param = 25,
  shape.param = numeric(0),
  cov.param = list(),
  truncation = 50
)
}
\arguments{
\item{key.function}{specifies shape of the detection function (either
half-normal "hn", hazard rate "hr" or uniform "uf")}

\item{scale.param}{numeric vector with either a single value to be applied globally or a value for each strata. These should be supplied on the natural scale.}

\item{shape.param}{numeric vector with either a single value to be applied globally or a value for each strata. These should be supplied on the natural scale.}

\item{cov.param}{Named list with one named entry per individual level covariate. Covariate parameter values should be defined on the log scale (rather than the natural scale), this is the same scale as provided in the ddf output in mrds and also in the MCDS output in Distance. Cluster sizes parameter values can be defined here. Each list entry will either be a data.frame containing 2 or 3 columns: level, param and where desired strata. If the region has multiple strata but this column is omitted then the values will be assumed to apply globally. The cluster size entry in the list must be named 'size'. Alternatively the list element may a numeric vector with either a single value to be applied globally or a value for each strata.}

\item{truncation}{the maximum perpendicular (or radial) distance at which
objects may be detected from a line (or point) transect.}
}
\value{
object of class Detectablility
}
\description{
The detectability of the population is described by the values in this
class.
}
\examples{
# The default values create a detectability object with a half normal
# detection function with scale parameter 25 and truncation distance 50.
}
\author{
Laura Marshall
}
