% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassConstructors.R
\name{make.ds.analysis}
\alias{make.ds.analysis}
\title{Creates an Analysis object}
\usage{
make.ds.analysis(
  dfmodel = list(~1),
  key = "hn",
  adjustment = list(),
  truncation = numeric(0),
  cutpoints = numeric(0),
  er.var = "R2",
  control.opts = list(),
  group.strata = data.frame(),
  criteria = "AIC"
)
}
\arguments{
\item{dfmodel}{list of distance sampling model formula specifying the detection function
(see \code{?ds} for further details)}

\item{key}{key function to use; "hn" gives half-normal (default) and "hr" gives
hazard-rate.}

\item{adjustment}{a way of providing information about the adjustment term options. A
list of options for adjustment parameters. In the case of multiple models this should
be a list of option lists, one for each model. Note adjustment terms can only be
included when there are no covariates in the model. The adjustment options include:
adjustment - "cos" (recommended), "herm" or "poly", order - the order of the adjustment
terms and scale - either "width" or "scale". See details for more information.}

\item{truncation}{distance can be supplied as (numeric, e.g. 5) or percentage (as a
string, e.g. "15\%"). By default for exact distances the maximum observed
distance is used as the right truncation. Note that any value supplied as a string
will be interpreted as a \% even without the \% symbol.}

\item{cutpoints}{if the data are binned, this vector gives the cutpoints of the bins.
Ensure that the first element is 0 (or the left truncation distance) and the last
is the distance to the end of the furthest bin. (Default NULL, no binning.) Note
that if data has columns distbegin and distend then these will be used as bins if cutpoints is not specified. If both are specified, cutpoints has precedence.}

\item{er.var}{encounter rate variance estimator to use when abundance estimates are
required. Defaults to "R2" for line transects and "P3" for point transects. See dht2
for more information and if more complex options are required.}

\item{control.opts}{A list of control options: method - optimisation method,
initial.values - a list of names starting values, see mrds}

\item{group.strata}{Dataframe with two columns ("design.id" and "analysis.id"). The
former gives the strata names as defined in the design (i.e. the region object) the
second specifies how they should be grouped (into less strata) for the analyses}

\item{criteria}{character model selection criteria (AIC, AICc, BIC)}
}
\value{
an object of class DS.Analysis
}
\description{
This method creates an Analysis objects which describes a one or more
models to fit to the distance data. The simulation will fit each of these
models to the data generated in the simulation and select the model with
the minimum criteria value.
}
\details{

}
\examples{

# To incorporate model selection between a 'hn' and 'hr' model:
ds.analyses <- make.ds.analysis(dfmodel = ~1, key = "hn")

}
\seealso{
\code{ds} in \code{library(Distance)}
}
\author{
Laura Marshall
}
