% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassConstructors.R
\name{make.simulation}
\alias{make.simulation}
\title{Creates a Simulation object}
\usage{
make.simulation(
  reps = 10,
  design = make.design(),
  population.description = make.population.description(),
  detectability = make.detectability(),
  ds.analysis = make.ds.analysis()
)
}
\arguments{
\item{reps}{number of times the simulation should be repeated}

\item{design}{an object of class Survey.Design created by a call to
\link{make.design}}

\item{population.description}{an object of class Population.Description
created by a call to \link{make.population.description}}

\item{detectability}{and object of class Detectabolity created by a call to
\link{make.detectability}}

\item{ds.analysis}{a list of objects of class DS.Analysis created by
a call to\link{make.ds.analysis}}
}
\value{
object of class Simulation
}
\description{
This creates a simulation with all the information necessary for DSsim
to generate a population, create or read in transects, simulate the survey process
and fit detection functions and estimate density / abundance. This function can be
used by itself based on default values to create a simple line transect example, see
Examples below. To create more complex simulations it is advisable to define the
different parts of the simulation individually before grouping them together. See
the Arguments for links to the functions which make the definitions for the
individual simulation components.
}
\details{
The \code{make.simulation} function is now set up so that by
 default (with the exception of specifying point transects rather than
  line) it can run a simple simulation example. See examples.
}
\examples{
# A basic line transect simulation example
}
\author{
Laura Marshall
}
