% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey.R
\name{plot,Survey,Region-method}
\alias{plot,Survey,Region-method}
\alias{plot,Survey,ANY-method}
\title{plot}
\usage{
\S4method{plot}{Survey,Region}(x, y, ...)

\S4method{plot}{Survey,ANY}(x, y = NULL, ...)
}
\arguments{
\item{x}{object of class Survey}

\item{y}{object of class Region}

\item{...}{additional plotting parameters}
}
\value{
Generate 4 plots showing the survey population, transects (including covered areas), detections and a histogram of the detection distances. Plots inlude the survey region. Also invisibly returns a list of ggplot objects if the user would like to customise the plots.

Generate 2 plots showing the survey population, transects (including covered areas), detections and a histogram of the detection distances. Plots do not include survey region. Also invisibly returns a list of ggplot objects if the user would like to customise the plots.
}
\description{
Produces four plots of the survey: 1) Plots the transects inside the survey
region, 2) plots the population, 3) plots the transects, population and
detections 4) plots a histogram of the detection distances. Note that only
plots 3 & 4 are generated without the survey region if Region is omitted.
}
