% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{plot,Density,ANY-method}
\alias{plot,Density,ANY-method}
\alias{plot,Density,Region-method}
\title{Plot}
\usage{
\S4method{plot}{Density,ANY}(x, y, strata = "all", title = "")

\S4method{plot}{Density,Region}(x, y, strata = "all", title = "")
}
\arguments{
\item{x}{object of class Density}

\item{y}{object of class Region}

\item{strata}{the strata name or number to be plotted. By default
all strata will be plotted.}

\item{title}{plot title}
}
\value{
ggplot object

ggplot object
}
\description{
Plots an S4 object of class 'Density'

Plots an S4 object of class 'Density'
}
