% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qztpois.R
\name{qztpois}
\alias{qztpois}
\title{Calculates quantile values from a zero-truncated Poisson distribution}
\usage{
qztpois(p, mean = NA)
}
\arguments{
\item{p}{values for which to calculate quantile values}

\item{mean}{mean of the distribution of interest}
}
\value{
returns quantile values from a zero-truncated Poisson distribution.
}
\description{
Calculates quantiles from a zero-truncated Poisson distribution with mean
equal to that specified. It uses an optimization routine to check which 
value of lambda will give values with the requested mean.
}
\note{
Internal function not intended to be called by user.
}
\author{
Len Thomas
}
