% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qztruncnbinom.R
\name{qztruncnbinom}
\alias{qztruncnbinom}
\title{Quantile function for a zero-truncated Negative Binomial distribution}
\usage{
qztruncnbinom(p, mean = NA, sd = NA)
}
\arguments{
\item{p}{vector of probabilities}

\item{mean}{target mean}

\item{sd}{target standard deviation}
}
\value{
vector of quantiles
}
\description{
Quantile function for a zero-truncated Negative Binomial distribution
}
