% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qztruncpoislognormal.R
\name{qztruncpoislognormal}
\alias{qztruncpoislognormal}
\title{Function to evaluate the quantiles of values p from a zero-truncated 
poisson log-normal distribution}
\usage{
qztruncpoislognormal(p, mean = NA, sd = NA, n_sim = 1e+05)
}
\arguments{
\item{p}{values for which to find quantiles}

\item{mean}{target mean of the generated values}

\item{sd}{target standard deviation of the generated values}

\item{n_sim}{number of simulations to use for quantile calculation}
}
\value{
vector of quantiles
}
\description{
Function to evaluate the quantiles of values p from a zero-truncated 
poisson log-normal distribution
}
\note{
Internal function not intended to be called by user.
}
\author{
Jack Nowacek
}
