% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.simulation.R
\name{run.simulation}
\alias{run.simulation}
\title{Method to run a simulation}
\usage{
run.simulation(
  simulation,
  run.parallel = FALSE,
  max.cores = NA,
  counter = TRUE,
  ...
)
}
\arguments{
\item{simulation}{\code{\link{Simulation-class}} object}

\item{run.parallel}{logical option to use multiple processors}

\item{max.cores}{integer maximum number of cores to use, if not specified then
one less than the number available will be used.}

\item{counter}{logical indicates if you would like to see the progress counter.}

\item{...}{minimum.n - specifying the minimum number of detections which are 
required before a detection function is fitted (defaults to 20), see details.}
}
\value{
the \code{\link{Simulation-class}} object which now includes
the results
}
\description{
Runs the simulation and returns the simulation object with results. If
running in parallel and max.cores is not specified it will default to using
one less than the number of cores / threads on your machine. For example
code see \code{\link{make.simulation}}
}
\details{
The minimum.n argument allows the user to reduce or increase the number 
of required detections for a detection function to be fitted and the
estimates of abundance and density obtained. This is implemented to make 
the simulation mode stable. We usually advise 60-80 detections when fitting
a detection function but by default we allow detetection functions to be 
fitted with as few as 20 detections. Please note that if a number of 
iterations are excluded based on this criteria then this will likely 
introduce bias into the estimates. Reducing the minimum.n value should
correct this bias but may lead to unstable model fitting within the
simulation.
}
\seealso{
\code{\link{make.simulation}}
}
