% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rztnbinom.R
\name{rztnbinom}
\alias{rztnbinom}
\title{Generates values from a zero-truncated Negative Binomial distribution with 
mean and variance equal to that specified. 
Function uses an optimization routine to check which values of parameters N and p 
will parametrize a distribution whose mean and variance 
match the specified values.}
\usage{
rztnbinom(n, mean = NA, sd = NA)
}
\arguments{
\item{n}{number of values to randomly generate}

\item{mean}{target mean of the generated values}

\item{sd}{target standard deviation of the generated values}
}
\value{
returns a vector of randomly generated values from a 
zero-truncated Negative Binomial distribution.
}
\description{
Generates values from a zero-truncated Negative Binomial distribution with 
mean and variance equal to that specified. 
Function uses an optimization routine to check which values of parameters N and p 
will parametrize a distribution whose mean and variance 
match the specified values.
}
\note{
Internal function not intended to be called by user.
}
\author{
Jack Nowacek
}
