% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rztpois.R
\name{rztpois}
\alias{rztpois}
\title{Randomly generates values from a zero-truncated Poisson distribution}
\usage{
rztpois(n, mean = NA)
}
\arguments{
\item{n}{number of values to randomly generate}

\item{mean}{mean of the generated values}
}
\value{
returns a randomly generated value from a zero-truncated Poisson
distribution.
}
\description{
Generates values from a zero-truncated Poisson distribution with mean
equal to that specified. It uses an optimisation routine to check which 
value of lambda will give values with the requested mean.
}
\note{
Internal function not intended to be called by user.
}
\author{
Len Thomas
}
