% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rztpoislognormal.R
\name{rztpoislognormal}
\alias{rztpoislognormal}
\title{Generates values from a zero-truncated Poisson Log-normal distribution with 
mean and variance equal to that specified. 
It uses an optimization routine to check which values of the parameters 
mu and sigma will parametrize a distribution whose mean and variance match
the specified values.}
\usage{
rztpoislognormal(n, mean = NA, sd = NA, verbose = FALSE)
}
\arguments{
\item{n}{number of values to randomly generate}

\item{mean}{target mean of the generated values}

\item{sd}{target standard deviation of the generated values}

\item{verbose}{boolean to print optimization details}
}
\value{
vector of randomly generated values
}
\description{
Generates values from a zero-truncated Poisson Log-normal distribution with 
mean and variance equal to that specified. 
It uses an optimization routine to check which values of the parameters 
mu and sigma will parametrize a distribution whose mean and variance match
the specified values.
}
\note{
Internal function not intended to be called by user.
}
\author{
Jack Nowacek
}
