% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{set.densities}
\alias{set.densities}
\title{Method to set density values}
\usage{
set.densities(density, densities)
}
\arguments{
\item{density}{object of class Density}

\item{densities}{a numeric vector of density values to update the
density grid with.}
}
\value{
returns the Density object with updated density values
}
\description{
This method sets the density values in a density object.
}
