% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.Summary.R
\name{show,Density.Summary-method}
\alias{show,Density.Summary-method}
\title{show}
\usage{
\S4method{show}{Density.Summary}(object)
}
\arguments{
\item{object}{object of class Density.Summary}
}
\value{
No return value, displays the density summary
}
\description{
displays the density summary table
}
