% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.Summary.R
\name{show,Simulation.Summary-method}
\alias{show,Simulation.Summary-method}
\title{show}
\usage{
\S4method{show}{Simulation.Summary}(object)
}
\arguments{
\item{object}{object of class Simulation.Summary}
}
\value{
No return value, displays information in Simulation.Summary
object
}
\description{
Displays the simulation summary
}
