% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{summary,Density-method}
\alias{summary,Density-method}
\title{summary}
\usage{
\S4method{summary}{Density}(object, ...)
}
\arguments{
\item{object}{object of class Simulation}

\item{...}{not implemented}
}
\value{
a \code{\link{Density.Summary-class}} object
}
\description{
Provides a summary table of the density object.
}
