% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{summary,Simulation-method}
\alias{summary,Simulation-method}
\title{summary}
\usage{
\S4method{summary}{Simulation}(object, description.summary = TRUE, use.max.reps = FALSE, ...)
}
\arguments{
\item{object}{object of class Simulation}

\item{description.summary}{logical indicating whether an
explanation of the summary should be displayed}

\item{use.max.reps}{by default this is FALSE meaning that only simulation repetitions where all models converged for that data set are included. By setting this to TRUE any repetition where one or more models converged will be included in the summary results.}

\item{...}{no additional arguments currently implemented}
}
\value{
Object of class Simulation.Summary
}
\description{
Provides a summary of the simulation results.
}
