% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey.LT.R
\docType{class}
\name{Survey.LT-class}
\alias{Survey.LT-class}
\title{Class "Survey.LT" extends class "Survey"}
\description{
Class \code{"Survey.LT"} is an S4 class containing a population
and a set of transects.
}
\section{Slots}{

\describe{
\item{\code{transect}}{Object of class \code{"Line.Transect"}; the
line transects.}

\item{\code{perpendicular.truncation}}{Object of class \code{"numeric"}; the
maximum distance from the transect at which animals may be detected.}
}}

\seealso{
\code{\link{make.design}}
}
\keyword{classes}
