% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey.PT.R
\docType{class}
\name{Survey.PT-class}
\alias{Survey.PT-class}
\title{Class "Survey.PT" extends class "Survey"}
\description{
Class \code{"Survey.PT"} is an S4 class containing a population
and a set of transects.
}
\section{Slots}{

\describe{
\item{\code{transect}}{Object of class \code{"Point.Transect"}; the
point transects.}

\item{\code{radial.truncation}}{Object of class \code{"numeric"}; the
maximum distance from the transect at which animals may be detected.}
}}

\seealso{
\code{\link{make.design}}
}
\keyword{classes}
