% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.functions.R, R/DS.Analysis.R
\name{analyse.data}
\alias{analyse.data}
\alias{analyse.data,DS.Analysis,Survey-method}
\alias{analyse.data,DS.Analysis,data.frame-method}
\title{S4 generic method to run analyses}
\usage{
analyse.data(analysis, data.obj, ...)

\S4method{analyse.data}{DS.Analysis,Survey}(analysis, data.obj, warnings = list(), ...)

\S4method{analyse.data}{DS.Analysis,data.frame}(analysis, data.obj, warnings = list(), transect = "line", ...)
}
\arguments{
\item{analysis}{an object of class DS.Analysis}

\item{data.obj}{an object of class Survey or a dataframe}

\item{...}{optional arguments (currently not used)}

\item{warnings}{a list of warnings and how many times they arose}

\item{transect}{character value either "line" or "point" specifying type of
transect used in survey}
}
\value{
a list containing an S3 ddf object and optionally an S3 dht object relating to the model with the minimum criteria.
}
\description{
This method carries out an analysis of distance sampling data. This method
is provided to allow the user to perform diagnostics of the analyses used
in the simulation. The data argument can be obtained by a call to
\code{simulate.survey(object, dht.table = TRUE)}. Note if the first object
supplied is of class DS.Analysis then the second argument must be of class
DDf.Data. The data argument may be of either class for an object argument
of class Simulation.
}
