% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.functions.R
\name{get.distance.data}
\alias{get.distance.data}
\title{S4 generic method to extract distance data}
\usage{
get.distance.data(object)
}
\arguments{
\item{object}{an object of class LT.Survey.Results}
}
\value{
a data.frame describing the distance data
}
\description{
Extracts distance data from a Survey.Results object
}
\seealso{
\code{\link{create.survey.results}}
}
