% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{histogram.N.ests}
\alias{histogram.N.ests}
\title{histogram.N.ests}
\usage{
histogram.N.ests(x, use.max.reps = FALSE, N.ests = "individuals", ...)
}
\arguments{
\item{x}{object of class Simulation}

\item{use.max.reps}{by default this is FALSE meaning that only simulation repetitions where all models converged for that data set are included. By setting this to TRUE any repetition where one or more models converged will be included in the summary results.}

\item{N.ests}{character indicating whether to plot estimates of abundance of 'individuals',
'clusters' or 'both'. By default this is individuals.}

\item{...}{optional parameters to pass to the generic hist function in graphics}
}
\value{
No return value, displays a histogram of the abundance estimates
}
\description{
Plots a histogram of the estimates abundances
}
