% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassConstructors.R
\name{make.population.description}
\alias{make.population.description}
\title{Creates a Population.Description object}
\usage{
make.population.description(
  region = make.region(),
  density = make.density(),
  covariates = list(),
  N = numeric(0),
  fixed.N = TRUE
)
}
\arguments{
\item{region}{the Region object in which this population exists (see \link{make.region}).}

\item{density}{the Density object describing the distribution of the individuals / clusters (see \link{make.density}).}

\item{covariates}{Named list with one named entry per individual level covariate. Cluster sizes can be defined here, it must be named 'size'. Each list entry should be another list with either one element or one element per strata allowing different population structures per strata. Each element of these lists should either be a data.frame containing 2 columns, the first the level (level) and the second the probability (prob). The cluster size entry in the list must be named 'size'. Alternatively the list element may be another list specifying the distribution in the first element and a named list in the second element with the distribution parameter.}

\item{N}{the number of individuals / clusters in a population (1000 by default)}

\item{fixed.N}{a logical value. If TRUE the population is generated from the value of N
otherwise it is generated from the density description.}
}
\value{
object of class Population.Description
}
\description{
Creates an object which describes a population. The values in this object
will be used to create instances of the population.
}
\details{
The \code{covariates} argument should specify a list with one named
element per covariate. The population covariate values can either be described
by dataframes representing discrete covariates or by distributions.
If specifying the covariate values via a distribution
this should be done in the form of a list, the list should first contain the
distribution and then specify

The first element should be one of
the following: 'normal', 'poisson', 'ztruncpois' or 'lognormal'. The 'ztruncpois'
distribution refers to a zero truncated Poisson distribution. The corresponding
parameters that you must supply are detailed below. These should be added to a named
list (each element named with the parameter name) containing the parameter values.
See examples for implementation.

\tabular{lll}{ Distribution  \tab Parameters  \tab         \cr
               normal        \tab mean        \tab sd      \cr
               poisson       \tab lambda      \tab         \cr
               ztruncpois    \tab mean        \tab         \cr
               lognormal     \tab meanlog     \tab sdlog   \cr
              }
}
\examples{
# Create population description example
}
\seealso{
\code{\link{make.region}}, \code{\link{make.density}}, \code{\link{make.detectability}}
}
\author{
Laura Marshall
}
