% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{plot,Density,ANY-method}
\alias{plot,Density,ANY-method}
\alias{plot,Density,Region-method}
\title{Plot}
\usage{
\S4method{plot}{Density,ANY}(x, y, strata = "all", title = "", scale = 1)

\S4method{plot}{Density,Region}(x, y, strata = "all", title = "", scale = 1, line.col = gray(0.2))
}
\arguments{
\item{x}{object of class Density}

\item{y}{object of class Region}

\item{strata}{the strata name or number to be plotted. By default
all strata will be plotted.}

\item{title}{plot title}

\item{scale}{used to scale the x and y values in the plot (warning may give
unstable results when a projection is defined for the study area!)}

\item{line.col}{sets the line colour for the shapefile}
}
\value{
ggplot object

ggplot object
}
\description{
Plots an S4 object of class 'Density'

Plots an S4 object of class 'Density'
}
