% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.simulation.R
\name{run.simulation}
\alias{run.simulation}
\title{Method to run a simulation}
\usage{
run.simulation(
  simulation,
  run.parallel = FALSE,
  max.cores = NA,
  save.data = FALSE,
  load.data = FALSE,
  data.path = character(),
  counter = TRUE,
  progress.file = character(),
  ...
)
}
\arguments{
\item{simulation}{an object of class Simulation}

\item{run.parallel}{logical option to use multiple processors}

\item{max.cores}{integer maximum number of cores to use, if not specified then
one less than the number available will be used.}

\item{save.data}{logical allows the datasets from the simulation to be
saved to file}

\item{load.data}{logical allows the datasets to be loaded from file rather than
simulated afresh.}

\item{data.path}{character file path to the data files.}

\item{counter}{logical indicates if you would like to see the progress counter.}

\item{progress.file}{character file to output progress to for Distance for Windows}

\item{...}{allows the five previous optional arguments to be specified}
}
\value{
an object of class simulation which now includes the results
}
\description{
Runs the simulation and returns the simulation object with results. If
running in parallel and max.cores is not specified it will default to using
one less than the number of cores / threads on your machine.
}
\seealso{
\code{\link{make.simulation}}
}
