% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.functions.R, R/Simulation.R,
%   R/Survey.LT.R, R/Survey.PT.R
\name{run.survey}
\alias{run.survey}
\alias{run.survey,Simulation-method}
\alias{run.survey,Survey.LT-method}
\alias{run.survey,Survey.PT-method}
\title{S4 generic method to simulate a survey}
\usage{
run.survey(object, ...)

\S4method{run.survey}{Simulation}(object)

\S4method{run.survey}{Survey.LT}(object, region = NULL)

\S4method{run.survey}{Survey.PT}(object, region = NULL)
}
\arguments{
\item{object}{an object of class Simulation}

\item{...}{allows a region object to be passed in}
}
\value{
an object of class LT.Survey.Results
}
\description{
Simulates the process by which individuals / clusters are detected.
Currently this is only implemented for line transect surveys. It returns an
object of class LT.Survey.Results which contains a population, a set of
transects, distance data and if requested region, sample and obs tables.
}
\examples{
\dontrun{
survey <- run.survey(simulation)

plot(survey)
}

}
